Fixed gravity when holding objects
Added exclude player layer on object colliders so they don't collide when the player is carrying them.

Also removed interactible objects from the ground layer to fix the super jump bug when holding a cube. This means we can't use jumping on cubes as a puzzle mechanic anymore (for now)