Bug found.
I found a bug while looking at BigLevel. It seems that, not only does the gravity wall seemingly only work properly when it feels like doing so (It's been pulling me to the side at times, then all of the sudden, in the same playtest, it works like it's supposed to, even allowing me to jump (Albeit getting me almost nowhere off the ground). Additionally, I feel like I need to do something about the walls; I tried using the conveyer walls, but I literally clipped outside of the level when I tried to transition from the vertical one to the horizontal one.