I created a new version of the Item Spawner.
If I'm going to be able to get this Conveyor Cannon level working, the regular Item Spawner won't work, since it can only have one item spawned at a time. Therefore, in this commit, I created a new version that can spawn in multiple instances of an item for the level.